import delete art component
