+++ src/Components/UploadedArtComponent/UploadedArt.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from "react";
 import { axiosWithBase } from "../../AxiosCustom";
 import { connect } from "react-redux";
 import Modal from "react-modal";
-import styled from "styled-components";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import EditForm from "../Test";
@@ -13,6 +12,7 @@ import {
   SellingSection
 } from "../../Views/BuyerOrderItems/BuyerOrderItemsStyle";
 import Spinner from "../Spinner";
+import ConfirmDelete from "./ConfirmDelete";
 
 const customStyles = {
   content: {
@@ -33,7 +33,7 @@ function ArtForSale(props) {
   const [artForSale, setArtForSale] = useState(null);
   const [modalIsOpen, setIsOpen] = React.useState(false);
   const [spinner, setSpinning] = useState(true);
-  const [deleteArtId, setDeleteArtId] = useState(null);
+  const [artToDelete, setArtToDelete] = useState(null);
 
   function openModal() {
     setIsOpen(true);
@@ -41,18 +41,6 @@ function ArtForSale(props) {
 
   function closeModal() {
     setIsOpen(false);
-  }
-
-  function deleteArt(id) {
-    axiosWithBase()
-      .delete(`/art/selling/${id}`)
-      .then(res => {
-        const updatedArt = props.artForSale.filter(art => art._id !== res._id);
-        props.setArtForSale(updatedArt);
-      })
-      .catch(() => {
-        toast.error("cannot delete art");
-      });
   }
 
   const id = props.loggedInUser._id;
@@ -96,21 +84,17 @@ function ArtForSale(props) {
                       <button
                         className="edit"
                         onClick={() => {
-                          setDeleteArtId(null);
+                          setArtToDelete(null);
                           openModal();
                         }}
                       >
                         Edit
                       </button>
                       <button
-                        onClick={
-                          () => {
-                            setDeleteArtId(art);
-                            openModal();
-                          }
-                          // setConfirmDelete(art._id)
-                          // props.history.push(`/selling/forsale/delete/${art._id}`);
-                        }
+                        onClick={() => {
+                          setArtToDelete(art);
+                          openModal();
+                        }}
                         className="delete"
                       >
                         Delete
@@ -123,26 +107,11 @@ function ArtForSale(props) {
                   onRequestClose={closeModal}
                   style={customStyles}
                 >
-                  {deleteArtId ? (
-                    <ConfirmContainer>
-                      <div className="confirm">
-                        <h2>Delete '{deleteArtId.name}'</h2>
-                        <div>
-                          <button
-                            className="cancel"
-                            onClick={() => setIsOpen(false)}
-                          >
-                            Cancel
-                          </button>
-                          <button
-                            className="delete"
-                            onClick={() => deleteArt(id)}
-                          >
-                            Delete
-                          </button>
-                        </div>
-                      </div>
-                    </ConfirmContainer>
+                  {artToDelete ? (
+                    <ConfirmDelete
+                      setIsOpen={setIsOpen}
+                      artToDelete={artToDelete}
+                    />
                   ) : (
                     <EditForm editArt={art} onRequestClose={closeModal} />
                   )}
@@ -175,58 +144,5 @@ function ArtForSale(props) {
     </MainContainer2>
   );
 }
-
-const ConfirmContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    background-color: white;
-    padding: 5px;
-
-    h2 {
-        font-size: 2rem
-    }
-
-    button {
-      color: white;
-      margin: 0 10px;
-      padding: 0.6rem 1.6rem;
-      font-size: 1.5rem;
-      border-radius: 5px;
-  }
-
-    .confirm {
-        display: flex;
-        flex-direction: column;
-        justify-content: space-evenly;
-        align-items: center;
-        height: 150px;
-        width: 300px;
-        padding: 1rem
-        border: 1px solid grey;
-        border-radius: 5px;
-        // background-color: yellow;
-      }
-      .cancel {
-        background-color: green;
-        border: 1px solid green;
-        &:hover {
-          color: green;
-          background: white;
-        }
-      }
-      
-      .delete {
-        background-color: red;
-        border: 1px solid red;
-        &:hover {
-          color: red;
-          background: white;
-        }
-      }
-      
-`;
 
 export default connect(state => state)(ArtForSale);
