+++ src/Views/BuyerOrderItems/BuyerOrderItemsStyle.js
@@ -58,7 +58,7 @@ export const StyledButton = styled(NavLink)`
 export const StyledOrderContainer = styled.div`
   width: 100%;
   height: 90%;
-  overflow-y: scroll;
+  // overflow-y: scroll;
 
   .nothing {
     width: 100%;
